+++ MSP App/API/Controllers/BackupController.cs
@@ -1,10 +1,10 @@
-﻿using API.Services;
-using Azure.Core;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
-using SharedComponents.Entities;
-using SharedComponents.WebEntities.Requests.BackupRequests;
+﻿using Microsoft.AspNetCore.Mvc;
+using SharedComponents.Entities.DbEntities;
+using SharedComponents.Entities.WebEntities.Requests.BackupRequests;
+using SharedComponents.Handlers.Attributes.HasPermission;
+using SharedComponents.Handlers.Results;
+using SharedComponents.Services.APIServices.Interfaces;
+using SharedComponents.Services.DbServices.Interfaces;
 
 namespace API.Controllers
 {
@@ -13,18 +13,22 @@ namespace API.Controllers
     [ApiExplorerSettings(GroupName = "v1-Web")]
     public class BackupController : ControllerBase
     {
-        private readonly DbBackupService _dbBackupService;
-        private readonly DbDeviceService _dbDeviceService;
-        private readonly DbNASServerService _dbNASServerService;
+        private readonly IDbBackupService _dbBackupService;
+        private readonly IDbDeviceService _dbDeviceService;
+        private readonly IDbNASServerService _dbNASServerService;
+        private readonly IAPIAuthService _authService;
 
-        public BackupController(DbBackupService dbBackupService, DbDeviceService dbDeviceService, DbNASServerService dbNASServerService)
+        public BackupController(IDbBackupService dbBackupService, IDbDeviceService dbDeviceService,
+            IDbNASServerService dbNASServerService, IAPIAuthService authService)
         {
             _dbBackupService = dbBackupService;
             _dbDeviceService = dbDeviceService;
             _dbNASServerService = dbNASServerService;
+            _authService = authService;
         }
 
         [HttpPost("")]
+        [HasPermission("Backup.Create.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> CreateAsync([FromBody] BackupCreateRequest request)
         {
             if (ModelState.IsValid)
@@ -34,6 +38,12 @@ namespace API.Controllers
                 {
                     return NotFound("Device not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 NASServer? nasServer = await _dbNASServerService.GetAsync(request.NASServerId);
                 if (nasServer == null)
                 {
@@ -64,6 +74,7 @@ namespace API.Controllers
         }
 
         [HttpPut("")]
+        [HasPermission("Backup.Update.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> UpdateAsync([FromBody] BackupUpdateRequest request)
         {
             if (ModelState.IsValid)
@@ -73,6 +84,12 @@ namespace API.Controllers
                 {
                     return NotFound("Backup not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), backup.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 backup.Filename = request.Name;
                 backup.Path = request.Path;
                 backup.Date = request.Date;
@@ -87,6 +104,7 @@ namespace API.Controllers
         }
 
         [HttpDelete("{id}")]
+        [HasPermission("Backup.Delete.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             if (ModelState.IsValid)
@@ -96,6 +114,12 @@ namespace API.Controllers
                 {
                     return NotFound("Backup not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), backup.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (await _dbBackupService.DeleteAsync(id))
                 {
                     return Ok("Backup deleted successfully.");
@@ -106,6 +130,7 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}")]
+        [HasPermission("Backup.Get.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAsync(int id)
         {
             if (ModelState.IsValid)
@@ -113,6 +138,12 @@ namespace API.Controllers
                 DeviceBackup? backup = await _dbBackupService.GetAsync(id);
                 if (backup != null)
                 {
+                    // Authentication check using roles + permissions
+                    if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), backup.TenantId))
+                    {
+                        return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                    }
+                    // --- End of authentication check ---
                     return Ok(backup);
                 }
                 return NotFound("Backup not found.");
@@ -121,16 +152,33 @@ namespace API.Controllers
         }
 
         [HttpGet("")]
+        [HasPermission("Backup.Get-All.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllAsync()
         {
             if (ModelState.IsValid)
             {
-                ICollection<DeviceBackup>? backups = await _dbBackupService.GetAllAsync();
+                var tenantIds = await _authService.GetUserAccessibleTenantsAsync(Request.Headers["Authorization"].ToString());
+                if (tenantIds == null)
+                {
+                    return new CustomForbidResult("User does not have any tenant permissions");
+                }
+                List<DeviceBackup>? backups = new List<DeviceBackup>();
+                foreach (var tenantId in tenantIds)
+                {
+                    if (tenantId != null)
+                    {
+                        var tenantBackups = await _dbBackupService.GetAllByTenantIdAsync((int)tenantId);
+                        if (tenantBackups != null)
+                        {
+                            backups.AddRange(tenantBackups);
+                        }
+                    }
+                }
                 if (backups != null)
                 {
                     if (backups.Any())
                     {
-                        return Ok(backups);
+                        return Ok(backups.Distinct());
                     }
                 }
                 return NotFound("No backups found.");
@@ -139,6 +187,7 @@ namespace API.Controllers
         }
 
         [HttpGet("By-Device/{deviceId}")]
+        [HasPermission("Backup.Get-By-Device.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllByDeviceIdAsync(int deviceId)
         {
             if (ModelState.IsValid)
@@ -148,6 +197,12 @@ namespace API.Controllers
                 {
                     return NotFound("Device not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 ICollection<DeviceBackup>? backups = await _dbBackupService.GetAllByClientIdAndUuidAsync(device.TenantId, device.DeviceInfo.ClientId, device.DeviceInfo.Uuid);
                 if (backups != null)
                 {
@@ -162,10 +217,17 @@ namespace API.Controllers
         }
 
         [HttpGet("By-Tenant/{tenantId}")]
+        [HasPermission("Backup.Get-By-Tenant.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllByTenantIdAsync(int tenantId)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<BackupController>(Request.Headers["Authorization"].ToString(), tenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 ICollection<DeviceBackup>? backups = await _dbBackupService.GetAllByTenantIdAsync(tenantId);
                 if (backups != null)
                 {
